pyproject.toml
[tool.poetry]
name = "delivery system"
version = "0.1.0"
description = "Implementation of the delivery system service as a monolith"
authors = ["Sandro <physrow@gmail.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^4.1.7"
djangorestframework = "^3.14.0"
psycopg2 = "^2.9.5"
pytest = "^7.2.1"
nox = "^2022.11.21"
django-phonenumber-field = "^7.0.2"
phonenumbers = "^8.13.6"
django-choices = "^1.7.2"
python-json-logger = "^2.0.7"
djangorestframework-simplejwt = "^5.2.2"


[tool.poetry.group.dev.dependencies]
ruff = "^0.0.252"
isort = "^5.12.0"
black = "^23.1.0"
mypy = "^1.0.1"
factory-boy = "^3.2.1"
pytest-django = "^4.5.2"
pytest-cov = "^4.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
